chore: make service aware of stored categories
Given that db should have only two categories, when starting the service
it inserts two new categories leading to a confussion about which are
the IDs is working with.

To solve this problem update category_service to check if is a cold
start meaning wether db has two categories or not, then based on the
result it will realize or populate that table.